requirements.txt
opencv-python
numpy
Pillow
gTTS
imutils
requests
picamera